@@ -11,15 +11,15 @@
     "url": "git+ssh://git@github.com/open-telemetry/opentelemetry-js.git"
   },
   "engines": {
-    "node": ">=14"
+    "node": "^18.19.0 || >=20.6.0"
   },
   "author": "OpenTelemetry Authors",
   "license": "Apache-2.0",
   "dependencies": {
     "@opentelemetry/api": "^1.3.0",
     "@opentelemetry/instrumentation-bunyan": "^0.34.0",
-    "@opentelemetry/resources": "^1.8.0",
-    "@opentelemetry/sdk-node": "^0.48.0",
+    "@opentelemetry/resources": "^2.0.0",
+    "@opentelemetry/sdk-node": "^0.200.0",
     "bunyan": "^1.8.15"
   }
 }
