@@ -19,7 +19,7 @@
     "tracing"
   ],
   "engines": {
-    "node": ">=14"
+    "node": "^18.19.0 || >=20.6.0"
   },
   "author": "OpenTelemetry Authors",
   "license": "Apache-2.0",
@@ -29,21 +29,21 @@
   "homepage": "https://github.com/open-telemetry/opentelemetry-js-contrib/tree/main/examples/express#readme",
   "dependencies": {
     "@opentelemetry/api": "^1.9.0",
-    "@opentelemetry/exporter-trace-otlp-proto": "^0.54.2",
-    "@opentelemetry/instrumentation": "^0.54.2",
-    "@opentelemetry/instrumentation-express": "^0.44.0",
-    "@opentelemetry/instrumentation-http": "^0.54.2",
-    "@opentelemetry/resources": "^1.27.0",
-    "@opentelemetry/sdk-trace-base": "^1.27.0",
-    "@opentelemetry/sdk-trace-node": "^1.27.0",
-    "@opentelemetry/semantic-conventions": "^1.27.0",
-    "axios": "^1.6.0",
-    "express": "^4.17.1"
+    "@opentelemetry/exporter-trace-otlp-proto": "^0.207.0",
+    "@opentelemetry/instrumentation": "^0.207.0",
+    "@opentelemetry/instrumentation-express": "^0.56.0",
+    "@opentelemetry/instrumentation-http": "^0.207.0",
+    "@opentelemetry/resources": "^2.2.0",
+    "@opentelemetry/sdk-trace-base": "^2.2.0",
+    "@opentelemetry/sdk-trace-node": "^2.2.0",
+    "@opentelemetry/semantic-conventions": "^1.37.0",
+    "axios": "^1.13.1",
+    "express": "^5.1.0"
   },
   "devDependencies": {
-    "@types/express": "^4.17.13",
+    "@types/express": "^5.0.5",
     "@types/node": "18.18.14",
-    "ts-node": "^10.6.0",
-    "typescript": "4.4.4"
+    "ts-node": "^10.9.2",
+    "typescript": "5.0.4"
   }
 }
