@@ -37,6 +37,7 @@
     "@opentelemetry/instrumentation": "^0.48.0",
     "@opentelemetry/instrumentation-http": "^0.48.0",
     "@opentelemetry/instrumentation-koa": "^0.31.0",
+    "@opentelemetry/resources": "^1.0.0",
     "@opentelemetry/sdk-trace-node": "^1.0.0",
     "@opentelemetry/sdk-trace-base": "^1.0.0",
     "@opentelemetry/semantic-conventions": "^1.27.0",
@@ -47,7 +48,7 @@
   "devDependencies": {
     "cross-env": "^7.0.0",
     "ts-node": "^10.6.0",
-    "typescript": "4.4.4",
+    "typescript": "5.0.4",
     "@types/koa": "^2.13.5"
   }
 }
