@@ -41,48 +44,57 @@
   "author": "OpenTelemetry Authors",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "5.8.1",
-    "@typescript-eslint/parser": "5.8.1",
-    "eslint": "8.7.0",
-    "eslint-config-airbnb-base": "15.0.0",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-header": "3.1.1",
-    "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "4.2.1",
-    "glob": "^10.3.10",
-    "lerna": "6.6.2",
-    "lerna-changelog": "2.2.0",
-    "markdownlint-cli2": "0.13.0",
-    "minimatch": "^9.0.3",
-    "nx": "15.9.7",
-    "mocha": "^10.7.3",
-    "prettier": "2.8.8",
+    "@babel/core": "7.28.5",
+    "@babel/preset-env": "7.28.5",
+    "@eslint/js": "9.34.0",
+    "@nx/eslint": "20.8.2",
+    "@types/mocha": "10.0.10",
+    "@types/node": "22.17.1",
+    "@types/semver": "7.7.1",
+    "@types/shimmer": "1.2.0",
+    "@types/sinon": "17.0.4",
+    "@types/webpack-env": "1.18.8",
+    "assert": "2.1.0",
+    "babel-loader": "10.0.0",
+    "babel-plugin-istanbul": "7.0.1",
+    "cross-env": "10.1.0",
+    "eslint": "9.34.0",
+    "eslint-plugin-import": "2.32.0",
+    "eslint-plugin-n": "17.21.3",
+    "eslint-plugin-yet-another-license-header": "0.2.0",
+    "expect": "29.7.0",
+    "glob": "11.0.3",
+    "globals": "^16.3.0",
+    "karma": "6.4.4",
+    "karma-chrome-launcher": "3.2.0",
+    "karma-coverage": "2.2.1",
+    "karma-mocha": "2.0.1",
+    "karma-spec-reporter": "0.0.36",
+    "karma-webpack": "5.0.1",
+    "lerna": "9.0.0",
+    "markdownlint-cli2": "0.17.2",
+    "minimatch": "9.0.5",
+    "mocha": "11.7.5",
+    "nock": "14.0.10",
+    "nx": "20.8.2",
+    "nyc": "17.1.0",
+    "prettier": "3.6.2",
     "process": "0.11.10",
-    "semver": "^7.6.0",
+    "rimraf": "5.0.10",
+    "safe-stable-stringify": "2.5.0",
+    "semver": "7.7.3",
+    "sinon": "15.2.0",
+    "test-all-versions": "6.2.0",
+    "ts-loader": "9.5.4",
     "ts-node": "10.9.2",
-    "util": "0.12.5"
-  },
-  "changelog": {
-    "labels": {
-      "breaking": ":boom: Breaking Change",
-      "bug": ":bug: Bug Fix",
-      "plugin-request": ":tada: New Plugins",
-      "enhancement": ":rocket: Enhancement",
-      "internal": ":house: Internal",
-      "documentation": ":memo: Documentation"
-    },
-    "ignoreCommitters": [
-      "renovate-bot",
-      "dependabot"
-    ]
+    "typescript": "5.0.4",
+    "typescript-eslint": "8.41.0",
+    "util": "0.12.5",
+    "webpack": "5.102.1",
+    "webpack-cli": "6.0.1",
+    "webpack-merge": "6.0.1"
   },
   "workspaces": [
-    "packages/*",
-    "plugins/node/*",
-    "plugins/web/*",
-    "propagators/*",
-    "detectors/node/*",
-    "metapackages/*"
+    "packages/*"
   ]
 }
