@@ -1,20 +1,22 @@
 {
   "name": "@opentelemetry/winston-transport",
-  "version": "0.10.1",
+  "version": "0.19.0",
   "description": "OpenTelemetry Transport for winston",
   "main": "build/src/index.js",
   "types": "build/src/index.d.ts",
-  "repository": "open-telemetry/opentelemetry-js-contrib",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/open-telemetry/opentelemetry-js-contrib.git",
+    "directory": "packages/winston-transport"
+  },
   "scripts": {
-    "test": "nyc mocha 'test/**/*.test.ts'",
-    "tdd": "npm run test -- --watch-extensions ts --watch",
     "clean": "rimraf build/*",
-    "lint": "eslint . --ext .ts",
-    "lint:fix": "eslint . --ext .ts --fix",
-    "prewatch": "npm run precompile",
+    "compile": "tsc -p .",
+    "compile:with-dependencies": "nx run-many -t compile -p @opentelemetry/winston-transport",
     "prepublishOnly": "npm run compile",
-    "version:update": "node ../../scripts/version-update.js",
-    "compile": "tsc -p ."
+    "tdd": "npm run test -- --watch-extensions ts --watch",
+    "test": "nyc mocha 'test/**/*.test.ts'",
+    "version:update": "node ../../scripts/version-update.js"
   },
   "keywords": [
     "transport",
@@ -26,7 +28,7 @@
   "author": "OpenTelemetry Authors",
   "license": "Apache-2.0",
   "engines": {
-    "node": ">=14"
+    "node": "^18.19.0 || >=20.6.0"
   },
   "files": [
     "build/src/**/*.js",
@@ -37,17 +39,10 @@
     "access": "public"
   },
   "devDependencies": {
-    "@types/mocha": "7.0.2",
-    "@types/node": "18.18.14",
-    "@types/sinon": "10.0.20",
-    "@types/triple-beam": "1.3.5",
-    "nyc": "15.1.0",
-    "rimraf": "5.0.10",
-    "sinon": "15.2.0",
-    "typescript": "4.4.4"
+    "@types/triple-beam": "1.3.5"
   },
   "dependencies": {
-    "@opentelemetry/api-logs": "^0.57.1",
+    "@opentelemetry/api-logs": "^0.208.0",
     "winston-transport": "4.*"
   },
   "homepage": "https://github.com/open-telemetry/opentelemetry-js-contrib/tree/main/packages/winston-transport#readme"
