@@ -1,14 +1,28 @@
-'use strict';
-
-import { setupTracing } from './tracer'
-const tracer = setupTracing('example-redis-server');
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 // Require in rest of modules
 import * as express from 'express';
-import axios from 'axios';
-import * as tracerHandlers from './express-tracer-handlers';
+import * as axios from 'axios';
 import { randomBytes } from 'crypto';
-const redisPromise = require('./setup-redis').redis;
+import { setupTracing } from './tracer';
+import * as tracerHandlers from './express-tracer-handlers';
+
+const tracer = setupTracing('example-redis-server');
+const { redisPromise } = require('./setup-redis');
 
 // Setup express
 const app = express();
@@ -32,7 +46,7 @@ async function setupRoutes() {
     }
   });
 
-  app.get('/:cmd', (req: any , res: any) => {
+  app.get('/:cmd', (req: any, res: any) => {
     if (!req.query.args) {
       res.status(400).send('No args provided');
       return;
@@ -54,8 +68,10 @@ async function setupRoutes() {
 
 // Setup express routes & middleware
 app.use(tracerHandlers.getMiddlewareTracer(tracer));
-setupRoutes().then(() => {
-  app.use(tracerHandlers.getErrorTracer(tracer));
-  app.listen(PORT);
-  console.log(`Listening on http://localhost:${PORT}`);
-});
+setupRoutes()
+  .then(() => {
+    app.use(tracerHandlers.getErrorTracer(tracer));
+    app.listen(PORT);
+    console.log(`Listening on http://localhost:${PORT}`);
+  })
+  .catch(err => console.log(err));
