@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+'use strict';
+
 // A small example that shows using OpenTelemetry's instrumentation of
 // Bunyan loggers. Usage:
 //    node --require ./telemetry.js app.js
@@ -21,9 +23,9 @@
 const otel = require('@opentelemetry/api');
 const bunyan = require('bunyan');
 
-const log = bunyan.createLogger({name: 'myapp', level: 'debug'});
+const log = bunyan.createLogger({ name: 'myapp', level: 'debug' });
 
-log.debug({foo: 'bar'}, 'hi');
+log.debug({ foo: 'bar' }, 'hi');
 
 const tracer = otel.trace.getTracer('example');
 tracer.startActiveSpan('manual-span', span => {
