@@ -21,13 +21,12 @@
 import { trace } from '@opentelemetry/api';
 import bunyan from 'bunyan';
 
-const log = bunyan.createLogger({name: 'myapp', level: 'debug'});
+const log = bunyan.createLogger({ name: 'myapp', level: 'debug' });
 
-log.debug({foo: 'bar'}, 'hi');
+log.debug({ foo: 'bar' }, 'hi');
 
 const tracer = trace.getTracer('example');
 tracer.startActiveSpan('manual-span', span => {
   log.info('this record will have trace_id et al fields for the current span');
   span.end();
 });
-
