@@ -1,6 +1,21 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
-// eslint-disable-next-line
 const tracing = require('./tracing')('example-connect-server');
 
 // Require in rest of modules
@@ -11,7 +26,6 @@ const axios = require('axios');
 const app = connect();
 const PORT = 8080;
 
-// eslint-disable-next-line prefer-arrow-callback
 app.use(function middleware1(req, res, next) {
   next();
 });
@@ -21,7 +35,9 @@ app.use((req, res, next) => {
 });
 
 app.use('/run_test', async (req, res) => {
-  const result = await axios.get('https://raw.githubusercontent.com/open-telemetry/opentelemetry-js/main/package.json');
+  const result = await axios.get(
+    'https://raw.githubusercontent.com/open-telemetry/opentelemetry-js/main/package.json'
+  );
   tracing.log('sending response');
   res.end(`OK ${result.data.version}`);
 
