@@ -1,24 +1,44 @@
-'use strict';
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-import { trace, SamplingDecision, SpanKind, Attributes } from '@opentelemetry/api';
+import { trace, SpanKind, Attributes } from '@opentelemetry/api';
 import { registerInstrumentations } from '@opentelemetry/instrumentation';
 import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
-import { Sampler, AlwaysOnSampler, SimpleSpanProcessor } from '@opentelemetry/sdk-trace-base';
+import {
+  Sampler,
+  AlwaysOnSampler,
+  SimpleSpanProcessor,
+  SamplingDecision,
+} from '@opentelemetry/sdk-trace-base';
 import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-proto';
-import { Resource } from '@opentelemetry/resources';
-import { ATTR_SERVICE_NAME, ATTR_HTTP_ROUTE } from '@opentelemetry/semantic-conventions';
+import {
+  ATTR_SERVICE_NAME,
+  ATTR_HTTP_ROUTE,
+} from '@opentelemetry/semantic-conventions';
 import { ExpressInstrumentation } from '@opentelemetry/instrumentation-express';
 import { HttpInstrumentation } from '@opentelemetry/instrumentation-http';
+import { resourceFromAttributes } from '@opentelemetry/resources';
 
 export const setupTracing = (serviceName: string) => {
   const exporter = new OTLPTraceExporter({});
   const provider = new NodeTracerProvider({
-    resource: new Resource({
+    resource: resourceFromAttributes({
       [ATTR_SERVICE_NAME]: serviceName,
     }),
-    spanProcessors: [
-      new SimpleSpanProcessor(exporter),
-    ],
+    spanProcessors: [new SimpleSpanProcessor(exporter)],
     sampler: filterSampler(ignoreHealthCheck, new AlwaysOnSampler()),
   });
   registerInstrumentations({
@@ -36,7 +56,11 @@ export const setupTracing = (serviceName: string) => {
   return trace.getTracer(serviceName);
 };
 
-type FilterFunction = (spanName: string, spanKind: SpanKind, attributes: Attributes) => boolean;
+type FilterFunction = (
+  spanName: string,
+  spanKind: SpanKind,
+  attributes: Attributes
+) => boolean;
 
 function filterSampler(filterFn: FilterFunction, parent: Sampler): Sampler {
   return {
@@ -48,10 +72,16 @@ function filterSampler(filterFn: FilterFunction, parent: Sampler): Sampler {
     },
     toString() {
       return `FilterSampler(${parent.toString()})`;
-    }
-  }
+    },
+  };
 }
 
-function ignoreHealthCheck(spanName: string, spanKind: SpanKind, attributes: Attributes) {
-  return spanKind !== SpanKind.SERVER || attributes[ATTR_HTTP_ROUTE] !== "/health";
+function ignoreHealthCheck(
+  spanName: string,
+  spanKind: SpanKind,
+  attributes: Attributes
+) {
+  return (
+    spanKind !== SpanKind.SERVER || attributes[ATTR_HTTP_ROUTE] !== '/health'
+  );
 }
