@@ -1,3 +1,19 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
 const url = require('url');
@@ -32,13 +48,13 @@ function makeRequest(query) {
         'Content-Type': 'application/json',
       },
     };
-    const req = http.request(options, (res) => {
+    const req = http.request(options, res => {
       const data = [];
-      res.on('data', (chunk) => data.push(chunk));
+      res.on('data', chunk => data.push(chunk));
       res.on('end', () => {
         resolve(data.toString());
       });
-      res.on('error', (err) => {
+      res.on('error', err => {
         reject(err);
       });
     });
