@@ -1,3 +1,19 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
 const api = require('@opentelemetry/api');
@@ -11,7 +27,7 @@ const { registerInstrumentations } = require('@opentelemetry/instrumentation');
 
 const EXPORTER = process.env.EXPORTER || '';
 
-module.exports = (serviceName) => {
+module.exports = serviceName => {
   let exporter;
   if (EXPORTER === 'jaeger') {
     exporter = new JaegerExporter({ serviceName });
@@ -20,9 +36,7 @@ module.exports = (serviceName) => {
   }
 
   const provider = new NodeTracerProvider({
-    spanProcessors: [
-      new SimpleSpanProcessor(exporter),
-    ],
+    spanProcessors: [new SimpleSpanProcessor(exporter)],
   });
 
   // Initialize the OpenTelemetry APIs to use the NodeTracerProvider bindings
