@@ -1,27 +1,41 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
 const opentelemetry = require('@opentelemetry/api');
 const { NodeTracerProvider } = require('@opentelemetry/sdk-trace-node');
 const { SimpleSpanProcessor } = require('@opentelemetry/sdk-trace-base');
 const { JaegerExporter } = require('@opentelemetry/exporter-jaeger');
-const { IORedisInstrumentation } = require('@opentelemetry/instrumentation-ioredis');
+const {
+  IORedisInstrumentation,
+} = require('@opentelemetry/instrumentation-ioredis');
 const { registerInstrumentations } = require('@opentelemetry/instrumentation');
 
 const exporter = new JaegerExporter({ serviceName: 'ioredis-example' });
 
 const provider = new NodeTracerProvider({
-  spanProcessors: [
-    new SimpleSpanProcessor(exporter),
-  ],
+  spanProcessors: [new SimpleSpanProcessor(exporter)],
 });
 
 // Initialize the OpenTelemetry APIs to use the BasicTracer bindings
 provider.register();
 
 registerInstrumentations({
-  instrumentations: [
-    new IORedisInstrumentation(),
-  ],
+  instrumentations: [new IORedisInstrumentation()],
   tracerProvider: provider,
 });
 
