@@ -1,13 +1,27 @@
-'use strict';
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 import * as api from '@opentelemetry/api';
-import { setupTracing } from './tracer'
-setupTracing('example-koa-server');
 
 // Adding Koa router (if desired)
-import * as Router from "@koa/router";
-import * as Koa from "koa"
+import * as Router from '@koa/router';
+import * as Koa from 'koa';
+import { setupTracing } from './tracer';
 
+setupTracing('example-koa-server');
 
 // Setup koa
 const app = new Koa();
@@ -15,7 +29,8 @@ const PORT = 8081;
 const router = new Router();
 
 // route definitions
-router.get('/run_test', runTest)
+router
+  .get('/run_test', runTest)
   .get('/post/new', addPost)
   .get('/post/:id', showNewPost);
 
@@ -26,7 +41,7 @@ async function setUp() {
 
 /**
  *  Router functions: list, add, or show posts
-*/
+ */
 const posts = ['post 0', 'post 1', 'post 2'];
 
 function addPost(ctx: Koa.Context) {
@@ -34,7 +49,7 @@ function addPost(ctx: Koa.Context) {
   posts.push(`post ${newPostId}`);
   const currentSpan = api.trace.getSpan(api.context.active());
   currentSpan?.addEvent('Added post');
-  currentSpan?.setAttribute('post.id', newPostId)
+  currentSpan?.setAttribute('post.id', newPostId);
   ctx.body = `Added post: ${posts[posts.length - 1]}`;
   ctx.redirect('/post/3');
 }
@@ -45,14 +60,16 @@ async function showNewPost(ctx: Koa.Context) {
   const post = posts[id];
   if (!post) ctx.throw(404, 'Invalid post id');
   const syntheticDelay = 500;
-  await new Promise((r) => setTimeout(r, syntheticDelay));
+  await new Promise(r => {
+    setTimeout(r, syntheticDelay);
+  });
   ctx.body = post;
 }
 
 function runTest(ctx: Koa.Context) {
   console.log('runTest');
   const currentSpan = api.trace.getSpan(api.context.active());
-  if (currentSpan){
+  if (currentSpan) {
     const { traceId } = currentSpan.spanContext();
     console.log(`traceid: ${traceId}`);
     console.log(`Jaeger URL: http://localhost:16686/trace/${traceId}`);
@@ -62,14 +79,15 @@ function runTest(ctx: Koa.Context) {
   }
 }
 
-async function noOp(ctx: Koa.Context, next: Koa.Next) {
+function noOp(ctx: Koa.Context, next: Koa.Next) {
   console.log('Sample basic koa middleware');
   const syntheticDelay = 100;
-  await new Promise((r) => setTimeout(r, syntheticDelay));
-  next();
+  setTimeout(next, syntheticDelay);
 }
 
-setUp().then(() => {
-  app.listen(PORT);
-  console.log(`Listening on http://localhost:${PORT}`);
-});
+setUp()
+  .then(() => {
+    app.listen(PORT);
+    console.log(`Listening on http://localhost:${PORT}`);
+  })
+  .catch(err => console.log(err));
