@@ -1,4 +1,18 @@
-'use strict';
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 import { KoaInstrumentation } from '@opentelemetry/instrumentation-koa';
 import { HttpInstrumentation } from '@opentelemetry/instrumentation-http';
@@ -10,7 +24,7 @@ import { SimpleSpanProcessor } from '@opentelemetry/sdk-trace-base';
 import { JaegerExporter } from '@opentelemetry/exporter-jaeger';
 import { ZipkinExporter } from '@opentelemetry/exporter-zipkin';
 import { Resource } from '@opentelemetry/resources';
-import { SEMRESATTRS_SERVICE_NAME } from '@opentelemetry/semantic-conventions'
+import { ATTR_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
 
 const EXPORTER = process.env.EXPORTER || '';
 
@@ -24,18 +38,13 @@ export const setupTracing = (serviceName: string) => {
 
   const provider = new NodeTracerProvider({
     resource: new Resource({
-      [SEMRESATTRS_SERVICE_NAME]: serviceName
+      [ATTR_SERVICE_NAME]: serviceName,
     }),
-    spanProcessors: [
-      new SimpleSpanProcessor(exporter),
-    ],
+    spanProcessors: [new SimpleSpanProcessor(exporter)],
   });
 
   registerInstrumentations({
-    instrumentations: [
-      new KoaInstrumentation(),
-      new HttpInstrumentation(),
-    ],
+    instrumentations: [new KoaInstrumentation(), new HttpInstrumentation()],
     tracerProvider: provider,
   });
 
