@@ -1,3 +1,19 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
 const opentelemetry = require('@opentelemetry/api');
@@ -7,27 +23,28 @@ diag.setLogger(new DiagConsoleLogger(), DiagLogLevel.VERBOSE);
 
 const { registerInstrumentations } = require('@opentelemetry/instrumentation');
 const { NodeTracerProvider } = require('@opentelemetry/sdk-trace-node');
-const { SimpleSpanProcessor, ConsoleSpanExporter } = require('@opentelemetry/sdk-trace-base');
+const {
+  SimpleSpanProcessor,
+  ConsoleSpanExporter,
+} = require('@opentelemetry/sdk-trace-base');
 const { Resource } = require('@opentelemetry/resources');
-const { SEMRESATTRS_SERVICE_NAME } = require('@opentelemetry/semantic-conventions');
+const { ATTR_SERVICE_NAME } = require('@opentelemetry/semantic-conventions');
 
-const { MemcachedInstrumentation } = require('@opentelemetry/instrumentation-memcached');
+const {
+  MemcachedInstrumentation,
+} = require('@opentelemetry/instrumentation-memcached');
 
-module.exports = (serviceName) => {
+module.exports = serviceName => {
   const exporter = new ConsoleSpanExporter();
   const provider = new NodeTracerProvider({
     resource: new Resource({
-      [SEMRESATTRS_SERVICE_NAME]: serviceName,
+      [ATTR_SERVICE_NAME]: serviceName,
     }),
-    spanProcessors: [
-      new SimpleSpanProcessor(exporter),
-    ],
+    spanProcessors: [new SimpleSpanProcessor(exporter)],
   });
   registerInstrumentations({
     tracerProvider: provider,
-    instrumentations: [
-      new MemcachedInstrumentation(),
-    ],
+    instrumentations: [new MemcachedInstrumentation()],
   });
 
   // Initialize the OpenTelemetry APIs to use the NodeTracerProvider bindings
