@@ -1,15 +1,28 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import * as api from '@opentelemetry/api';
-
-import { setupTracing } from './tracer';
-
-setupTracing('example-mongodb-server');
-
-import { accessDB } from './utils';
 
 import * as http from 'http';
 import { IncomingMessage, ServerResponse } from 'http';
 import * as mongodb from 'mongodb';
 import { Collection } from 'mongodb';
+import { setupTracing } from './tracer';
+import { accessDB } from './utils';
+
+setupTracing('example-mongodb-server');
 
 const DB_NAME = 'mydb';
 const COLLECTION_NAME = 'users';
@@ -41,7 +54,7 @@ function handleRequest(request: IncomingMessage, response: ServerResponse) {
   const currentSpan = api.trace.getSpan(api.context.active());
   if (currentSpan) {
     // display traceID in the terminal
-    const { traceId } = currentSpan?.spanContext();
+    const { traceId } = currentSpan.spanContext();
     console.log(`traceid: ${traceId}`);
     console.log(`Jaeger URL: http://localhost:16686/trace/${traceId}`);
     console.log(`Zipkin URL: http://localhost:9411/zipkin/traces/${traceId}`);
@@ -79,9 +92,12 @@ function handleInsertQuery(response: ServerResponse) {
     .then(() => {
       console.log('1 document inserted');
       // find document to test context
-      usersCollection.findOne({}).then(res => {
-        console.log(JSON.stringify(res));
-      });
+      usersCollection
+        .findOne({})
+        .then(res => {
+          console.log(JSON.stringify(res));
+        })
+        .catch(err => console.log(err));
     })
     .catch(err => {
       console.log('Error code:', err.code);
