@@ -1,25 +1,44 @@
-'use strict';
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 import opentelemetry from '@opentelemetry/api';
 import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
-import { SimpleSpanProcessor, SpanProcessor } from '@opentelemetry/sdk-trace-base';
+import {
+  SimpleSpanProcessor,
+  SpanProcessor,
+} from '@opentelemetry/sdk-trace-base';
 import { ZipkinExporter } from '@opentelemetry/exporter-zipkin';
 import { registerInstrumentations } from '@opentelemetry/instrumentation';
 import { HttpInstrumentation } from '@opentelemetry/instrumentation-http';
 import { MySQLInstrumentation } from '@opentelemetry/instrumentation-mysql';
 import { Resource } from '@opentelemetry/resources';
-import { SEMRESATTRS_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
+import { ATTR_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
 import {
   MeterProvider,
   PeriodicExportingMetricReader,
 } from '@opentelemetry/sdk-metrics';
-const { OTLPMetricExporter } = require('@opentelemetry/exporter-metrics-otlp-grpc');
+
+const {
+  OTLPMetricExporter,
+} = require('@opentelemetry/exporter-metrics-otlp-grpc');
 
 const EXPORTER = process.env.EXPORTER || '';
 
 export const setupTracing = (serviceName: string) => {
-
-  //metrics:
+  // metrics:
   const metricExporter = new OTLPMetricExporter();
   const metricReader = new PeriodicExportingMetricReader({
     exporter: metricExporter,
@@ -30,7 +49,7 @@ export const setupTracing = (serviceName: string) => {
     readers: [metricReader],
   });
 
-  //traces:
+  // traces:
   const spanProcessors: SpanProcessor[] = [];
 
   if (EXPORTER.toLowerCase().startsWith('z')) {
@@ -39,7 +58,7 @@ export const setupTracing = (serviceName: string) => {
 
   const tracerProvider = new NodeTracerProvider({
     resource: new Resource({
-      [SEMRESATTRS_SERVICE_NAME]: serviceName,
+      [ATTR_SERVICE_NAME]: serviceName,
     }),
     spanProcessors,
   });
@@ -48,12 +67,9 @@ export const setupTracing = (serviceName: string) => {
   tracerProvider.register();
 
   registerInstrumentations({
-    instrumentations: [
-      new HttpInstrumentation(),
-      new MySQLInstrumentation(),
-    ],
-    tracerProvider: tracerProvider,
-    meterProvider: meterProvider,
+    instrumentations: [new HttpInstrumentation(), new MySQLInstrumentation()],
+    tracerProvider,
+    meterProvider,
   });
 
   return opentelemetry.trace.getTracer('mysql-example');
