@@ -1,3 +1,19 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
 const { diag, DiagConsoleLogger, DiagLogLevel } = require('@opentelemetry/api');
@@ -6,14 +22,19 @@ const { HttpInstrumentation } = require('@opentelemetry/instrumentation-http');
 const { NetInstrumentation } = require('@opentelemetry/instrumentation-net');
 const { DnsInstrumentation } = require('@opentelemetry/instrumentation-dns');
 const { registerInstrumentations } = require('@opentelemetry/instrumentation');
-const { SimpleSpanProcessor, ConsoleSpanExporter } = require('@opentelemetry/sdk-trace-base');
+const {
+  SimpleSpanProcessor,
+  ConsoleSpanExporter,
+} = require('@opentelemetry/sdk-trace-base');
 const { JaegerExporter } = require('@opentelemetry/exporter-jaeger');
 
 const provider = new NodeTracerProvider({
   spanProcessors: [
-    new SimpleSpanProcessor(new JaegerExporter({
-      serviceName: 'http-client',
-    })),
+    new SimpleSpanProcessor(
+      new JaegerExporter({
+        serviceName: 'http-client',
+      })
+    ),
     new SimpleSpanProcessor(new ConsoleSpanExporter()),
   ],
 });
@@ -39,14 +60,20 @@ require('dns');
 const https = require('https');
 const http = require('http');
 
-http.get('http://opentelemetry.io/', () => {}).on('error', (e) => {
-  console.error(e);
-});
+http
+  .get('http://opentelemetry.io/', () => {})
+  .on('error', e => {
+    console.error(e);
+  });
 
-https.get('https://opentelemetry.io/', () => {}).on('error', (e) => {
-  console.error(e);
-});
+https
+  .get('https://opentelemetry.io/', () => {})
+  .on('error', e => {
+    console.error(e);
+  });
 
-https.get('https://opentelemetry.io/', { ca: [] }, () => {}).on('error', (e) => {
-  console.error(e);
-});
+https
+  .get('https://opentelemetry.io/', { ca: [] }, () => {})
+  .on('error', e => {
+    console.error(e);
+  });
