@@ -1,11 +1,32 @@
-import { ConsoleSpanExporter, SimpleSpanProcessor } from '@opentelemetry/sdk-trace-base';
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict';
+
+import {
+  ConsoleSpanExporter,
+  SimpleSpanProcessor,
+} from '@opentelemetry/sdk-trace-base';
 import { diag, DiagConsoleLogger, DiagLogLevel } from '@opentelemetry/api';
 import { WebTracerProvider } from '@opentelemetry/sdk-trace-web';
 import { BaseOpenTelemetryComponent } from '@opentelemetry/plugin-react-load';
 import { ZoneContextManager } from '@opentelemetry/context-zone';
 import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http';
 
-export default (serviceName) => {
+export default serviceName => {
   diag.setLogger(new DiagConsoleLogger(), DiagLogLevel.DEBUG);
   const exporter = new OTLPTraceExporter({
     url: 'http://localhost:4318/v1/traces',
@@ -23,7 +44,7 @@ export default (serviceName) => {
 
   const tracer = provider.getTracer(serviceName);
 
-  BaseOpenTelemetryComponent.setTracer(serviceName)
+  BaseOpenTelemetryComponent.setTracer(serviceName);
 
   return tracer;
-}
+};
