@@ -1,20 +1,41 @@
-import { ConsoleSpanExporter, SimpleSpanProcessor } from '@opentelemetry/sdk-trace-base';
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict';
+
+import {
+  ConsoleSpanExporter,
+  SimpleSpanProcessor,
+} from '@opentelemetry/sdk-trace-base';
 import { WebTracerProvider } from '@opentelemetry/sdk-trace-web';
 import { BaseOpenTelemetryComponent } from '@opentelemetry/plugin-react-load';
 import { ZoneContextManager } from '@opentelemetry/context-zone';
 import { CollectorTraceExporter } from '@opentelemetry/exporter-collector';
 import { diag, DiagConsoleLogger } from '@opentelemetry/api';
 import { Resource } from '@opentelemetry/resources';
-import { SEMRESATTRS_SERVICE_NAME } from '@opentelemetry/semantic-conventions'
+import { ATTR_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
 
-export default (serviceName) => {
+export default serviceName => {
   const exporter = new CollectorTraceExporter({
     url: 'http://localhost:55678/v1/trace',
   });
 
   const provider = new WebTracerProvider({
     resource: new Resource({
-        [SEMRESATTRS_SERVICE_NAME]: "react-load-example"
+      [ATTR_SERVICE_NAME]: 'react-load-example',
     }),
     spanProcessors: [
       new SimpleSpanProcessor(new ConsoleSpanExporter()),
@@ -28,8 +49,8 @@ export default (serviceName) => {
 
   const tracer = provider.getTracer(serviceName);
 
-  BaseOpenTelemetryComponent.setTracer(serviceName)
+  BaseOpenTelemetryComponent.setTracer(serviceName);
   diag.setLogger(new DiagConsoleLogger());
 
   return tracer;
-}
+};
