@@ -1,3 +1,19 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
 require('./tracer')('example-router-server');
@@ -16,24 +32,22 @@ const router = Router();
 router.use(setDefaultName);
 
 router.param('name', (req, res, next, name) => {
-  req.params.name = typeof name === 'string' ? name.toUpperCase() : req.defaultName;
+  req.params.name =
+    typeof name === 'string' ? name.toUpperCase() : req.defaultName;
   next();
 });
 
-// eslint-disable-next-line prefer-arrow-callback
 router.get('/hello/:name', function greetingHandler(req, res) {
   res.setHeader('Content-Type', 'text/plain; charset=utf-8');
   res.end(`Hello, ${req.params.name}!`);
 });
 
-// eslint-disable-next-line prefer-arrow-callback
 router.get('/err', function erroringRoute(req, res, next) {
   next(new Error('Broken!'));
 });
 
-// eslint-disable-next-line prefer-arrow-callback, func-names
 const server = http.createServer(function (req, res) {
-  router(req, res, (error) => {
+  router(req, res, error => {
     if (error) {
       res.statusCode = 500;
     } else {
