@@ -1,3 +1,19 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
 const webpack = require('webpack');
@@ -38,10 +54,7 @@ const common = {
     ],
   },
   resolve: {
-    modules: [
-      path.resolve(directory),
-      'node_modules',
-    ],
+    modules: [path.resolve(directory), 'node_modules'],
     extensions: ['.ts', '.js', '.jsx', '.json'],
   },
 };
@@ -49,7 +62,7 @@ const common = {
 module.exports = webpackMerge(common, {
   devtool: 'eval-source-map',
   devServer: {
-    static: path.resolve(path.join(__dirname, "examples")),
+    static: path.resolve(path.join(__dirname, 'examples')),
   },
   plugins: [
     new webpack.DefinePlugin({
