@@ -15,4 +15,5 @@
  */
 
 export { BaggageSpanProcessor } from './baggage-span-processor';
-export { ALLOW_ALL_BAGGAGE_KEYS, BaggageKeyPredicate } from './types';
+export { ALLOW_ALL_BAGGAGE_KEYS } from './types';
+export type { BaggageKeyPredicate } from './types';
