@@ -22,7 +22,7 @@
  * See "Rule:" comments for things that are checked.
  *
  * Usage:
- *      node scripts/lint-semconv-deps.js
+ *      node scripts/lint-semconv-deps.js [-x EXCLUDE-PATH]
  */
 
 import fs from 'fs';
@@ -52,12 +52,12 @@ function getAllWorkspaceDirs() {
     fs.readFileSync(path.join(TOP, 'package.json'), 'utf8')
   );
   return pj.workspaces
-    .map((wsGlob) => globSync(path.join(wsGlob, 'package.json')))
+    .map(wsGlob => globSync(path.join(wsGlob, 'package.json')))
     .flat()
     .map(path.dirname);
 }
 
-function lintSemconvDeps() {
+function lintSemconvDeps(excludePaths) {
   const wsDirs = getAllWorkspaceDirs();
 
   for (let wsDir of wsDirs) {
@@ -73,27 +73,45 @@ function lintSemconvDeps() {
     // Rule: The semconv dep should *not* be pinned. Expect `^X.Y.Z`.
     const pinnedVerRe = /^\d+\.\d+\.\d+$/;
     if (depRange && pinnedVerRe.exec(depRange)) {
-      problem(`${wsDir}/package.json: package ${pj.name} pins "${SEMCONV}" in dependencies, but should not (see https://github.com/open-telemetry/opentelemetry-js/tree/main/semantic-conventions#why-not-pin-the-version)`);
+      problem(
+        `${wsDir}/package.json: package ${pj.name} pins "${SEMCONV}" in dependencies, but should not (see https://github.com/open-telemetry/opentelemetry-js/tree/main/semantic-conventions#why-not-pin-the-version)`
+      );
     } else if (devDepRange && pinnedVerRe.exec(devDepRange)) {
-      problem(`${wsDir}/package.json: package ${pj.name} pins "${SEMCONV}" in devDependencies, but should not (see https://github.com/open-telemetry/opentelemetry-js/tree/main/semantic-conventions#why-not-pin-the-version)`);
+      problem(
+        `${wsDir}/package.json: package ${pj.name} pins "${SEMCONV}" in devDependencies, but should not (see https://github.com/open-telemetry/opentelemetry-js/tree/main/semantic-conventions#why-not-pin-the-version)`
+      );
     }
 
     // Rule: The incubating entry-point should not be used.
     const srcFiles = globSync(path.join(wsDir, 'src', '**', '*.ts'));
-    const usesIncubatingRe = /import\s+\{?[^{;]*\s+from\s+'@opentelemetry\/semantic-conventions\/incubating'/s;
+    const usesIncubatingRe =
+      /import\s+\{?[^{;]*\s+from\s+'@opentelemetry\/semantic-conventions\/incubating'/s;
     for (let srcFile of srcFiles) {
+      if (excludePaths.includes(srcFile)) {
+        continue;
+      }
       const srcText = fs.readFileSync(srcFile, 'utf8');
       const match = usesIncubatingRe.exec(srcText);
       if (match) {
-        problem(`${srcFile}: uses the 'incubating' entry-point from '@opentelemetry/semantic-conventions', but should not (see https://github.com/open-telemetry/opentelemetry-js/tree/main/semantic-conventions#unstable-semconv)`)
+        problem(
+          `${srcFile}: uses the 'incubating' entry-point from '@opentelemetry/semantic-conventions', but should not (see https://github.com/open-telemetry/opentelemetry-js/tree/main/semantic-conventions#unstable-semconv)`
+        );
       }
     }
   }
 }
 
-// mainline
-await lintSemconvDeps();
+// ---- mainline
+
+// Collect `-x EXCLUDE-PATH` args.
+const excludePaths = [];
+for (let i = 2; i < process.argv.length; i += 2) {
+  if (process.argv[i] === '-x' && process.argv.length > i + 1) {
+    excludePaths.push(process.argv[i + 1]);
+  }
+}
+
+await lintSemconvDeps(excludePaths);
 if (numProbs > 0) {
   process.exitCode = 1;
 }
-
