@@ -1,4 +1,4 @@
-/*!
+/*
  * Copyright The OpenTelemetry Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,7 +15,6 @@
  */
 
 const fs = require('fs');
-const os = require('os');
 const path = require('path');
 
 const appRoot = process.cwd();
@@ -44,6 +43,6 @@ export const PACKAGE_VERSION = '${pjson.version}';
 export const PACKAGE_NAME = '${pjson.name}';
 `;
 
-const fileUrl = path.join(appRoot, 'src', 'version.ts')
+const fileUrl = path.join(appRoot, 'src', 'version.ts');
 
 fs.writeFileSync(fileUrl, content);
